Move row_condition logic from Expectation to RendererConfiguration
