+++ great_expectations/expectations/expectation.py
@@ -1590,80 +1590,6 @@ class Expectation(metaclass=MetaExpectation):
         return value_set_str
 
     @staticmethod
-    def _get_row_conditions_list_from_row_condition_str(
-        row_condition_str: str,
-    ) -> List[str]:
-        # divide the whole condition into smaller parts
-        row_conditions_list = re.split(r"AND|OR|NOT(?! in)|\(|\)", row_condition_str)
-        row_conditions_list = [
-            condition.strip() for condition in row_conditions_list if condition.strip()
-        ]
-        return row_conditions_list
-
-    @staticmethod
-    def _parse_row_condition_str(row_condition_str: str) -> str:
-        if not row_condition_str:
-            row_condition_str = "True"
-
-        row_condition_str = (
-            row_condition_str.replace("&", " AND ")
-            .replace(" and ", " AND ")
-            .replace("|", " OR ")
-            .replace(" or ", " OR ")
-            .replace("~", " NOT ")
-            .replace(" not ", " NOT ")
-        )
-        row_condition_str = " ".join(row_condition_str.split())
-
-        # replace tuples of values by lists of values
-        tuples_list = re.findall(r"\([^()]*,[^()]*\)", row_condition_str)
-        for value_tuple in tuples_list:
-            value_list = value_tuple.replace("(", "[").replace(")", "]")
-            row_condition_str = row_condition_str.replace(value_tuple, value_list)
-
-        return row_condition_str
-
-    @staticmethod
-    @param_method(param_name="row_condition")
-    def _add_row_condition_params(
-        renderer_configuration: RendererConfiguration,
-    ) -> RendererConfiguration:
-        row_condition_str: str = Expectation._parse_row_condition_str(
-            row_condition_str=renderer_configuration.params.row_condition.value
-        )
-        row_conditions_list: List[
-            str
-        ] = Expectation._get_row_conditions_list_from_row_condition_str(
-            row_condition_str=row_condition_str
-        )
-        for idx, condition in enumerate(row_conditions_list):
-            name = f"row_condition__{str(idx)}"
-            value = condition.replace(" NOT ", " not ")
-            renderer_configuration.add_param(
-                name=name, schema_type=RendererSchemaType.STRING, value=value
-            )
-
-        return renderer_configuration
-
-    @staticmethod
-    @param_method(param_name="row_condition")
-    def _get_row_condition_string(renderer_configuration: RendererConfiguration) -> str:
-        row_condition_str: str = Expectation._parse_row_condition_str(
-            row_condition_str=renderer_configuration.params.row_condition.value
-        )
-        row_conditions_list: List[
-            str
-        ] = Expectation._get_row_conditions_list_from_row_condition_str(
-            row_condition_str=row_condition_str
-        )
-        for idx, condition in enumerate(row_conditions_list):
-            row_condition_str = row_condition_str.replace(
-                condition, f"$row_condition__{str(idx)}"
-            )
-        row_condition_str = row_condition_str.lower()
-        return f"if {row_condition_str}"
-
-    @staticmethod
     @param_method(param_name="mostly")
     def _add_mostly_pct_param(
         renderer_configuration: RendererConfiguration,
